backend/requirements.txt
Package                       Version
----------------------------- ------------
asgiref                       3.8.1
distlib                       0.3.8
Django                        4.1.13
django-cors-headers           4.4.0
django-environ                0.11.2
djangorestframework           3.15.2
djangorestframework-simplejwt 5.3.1
djongo                        1.3.6
dnspython                     2.6.1
filelock                      3.16.1
geographiclib                 2.0
geopy                         2.4.1
pip                           24.2
platformdirs                  4.3.6
PyJWT                         2.9.0
pymongo                       3.12.3
pypng                         0.20220715.0
pytz                          2024.2
qrcode                        7.4.2
sqlparse                      0.2.4
typing_extensions             4.12.2
virtualenv                    20.26.3
